Move AppHost heuristic to Projects utility
